all auth types working with some edge cases fixed
