+++ .gitignore
@@ -3,4 +3,3 @@ config.js
 auth.js
 gulpfile.babel.js
 compiled
-gulpfile.babel.js
