+++ db/read_user_by_email.sql
@@ -1,0 +1,2 @@
+SELECT * FROM users
+WHERE email = $1;
