+++ db/update_google_id.sql
@@ -1,0 +1,2 @@
+UPDATE users SET google = $2
+WHERE email = $1;
