+++ db/update_user_password.sql
@@ -1,0 +1,2 @@
+UPDATE users SET phone = $2, password = $3
+WHERE email = $1;
