+++ public/compiled/styles.css
@@ -532,7 +532,8 @@ h1, h2, h3, h4, h5, h6 {
   display: none;
   position: fixed;
   overflow: hidden;
-  z-index: 4; }
+  z-index: 4;
+  font-family: 'Source Sans Pro'; }
   .menu .menu-box-container {
     margin-top: 10vh;
     height: 50vh;
